Add files via upload
The user should provide the path where to save the file and the starting year, ending year and the starting month, ending month, and the starting date and ending month. and then the the data are extracted into two columns excel sheet  for each date and the corresponding average wind speed.